refactor: step1 코드 패키지 이동
