+++ src/test/java/study/SetTest.java
@@ -1,4 +1,4 @@
-package study.car;
+package study;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
