+++ src/test/java/study/StringTest.java
@@ -1,4 +1,4 @@
-package study.car;
+package study;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
